+++ akka-cluster-sharding/src/test/java/akka/cluster/sharding/ClusterShardingTest.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2009-2015 Typesafe Inc. <http://www.typesafe.com>
  */
 
-package akka.contrib.pattern;
+package akka.cluster.sharding;
 
 import java.util.concurrent.TimeUnit;
