+++ akka-cluster-sharding/src/test/resources/reference.conf
@@ -1,0 +1,6 @@
+akka {
+  actor {
+    serialize-creators = on
+    serialize-messages = on
+  }
+}
