+++ akka-cluster-sharding/src/test/scala/akka/cluster/sharding/LeastShardAllocationStrategySpec.scala
@@ -1,7 +1,7 @@
 /**
  * Copyright (C) 2009-2015 Typesafe Inc. <http://www.typesafe.com>
  */
-package akka.contrib.pattern
+package akka.cluster.sharding
 
 import akka.testkit.AkkaSpec
 import akka.actor.Props
