+++ akka-docs/rst/java/index-network.rst
@@ -9,6 +9,7 @@ Networking
    ../scala/cluster-singleton
    ../scala/distributed-pub-sub
    ../scala/cluster-client
+   ../scala/cluster-sharding
    cluster-metrics
    remoting
    serialization
