+++ project/AkkaBuild.scala
@@ -55,8 +55,8 @@ object AkkaBuild extends Build {
         archivesPathFinder.get.map(file => (file -> ("akka/" + file.getName)))
       }
     ),
-    aggregate = Seq(actor, testkit, actorTests, remote, remoteTests, camel, cluster, clusterMetrics, clusterTools, slf4j, agent,
-      persistence, persistenceTck, kernel, osgi, docs, contrib, samples, multiNodeTestkit, typed)
+    aggregate = Seq(actor, testkit, actorTests, remote, remoteTests, camel, cluster, clusterMetrics, clusterTools, clusterSharding,
+      slf4j, agent, persistence, persistenceTck, kernel, osgi, docs, contrib, samples, multiNodeTestkit, typed)
   )
 
   lazy val akkaScalaNightly = Project(
@@ -64,8 +64,8 @@ object AkkaBuild extends Build {
     base = file("akka-scala-nightly"),
     // remove dependencies that we have to build ourselves (Scala STM)
     // samples don't work with dbuild right now
-    aggregate = Seq(actor, testkit, actorTests, remote, remoteTests, camel, cluster, clusterMetrics, clusterTools, slf4j,
-      persistence, persistenceTck, kernel, osgi, contrib, multiNodeTestkit, typed)
+    aggregate = Seq(actor, testkit, actorTests, remote, remoteTests, camel, cluster, clusterMetrics, clusterTools, clusterSharding,
+      slf4j, persistence, persistenceTck, kernel, osgi, contrib, multiNodeTestkit, typed)
   ).disablePlugins(ValidatePullRequest)
 
   lazy val actor = Project(
@@ -131,6 +131,13 @@ object AkkaBuild extends Build {
     id = "akka-cluster-tools",
     base = file("akka-cluster-tools"),
     dependencies = Seq(cluster % "compile->compile;test->test;multi-jvm->multi-jvm")
+  ) configs (MultiJvm)
+  
+  lazy val clusterSharding = Project(
+    id = "akka-cluster-sharding",
+    base = file("akka-cluster-sharding"),
+    dependencies = Seq(cluster % "compile->compile;test->test;multi-jvm->multi-jvm", 
+        persistence % "compile;test->provided", clusterTools)
   ) configs (MultiJvm)
 
   lazy val slf4j = Project(
